+++ lib/flutter_branch_sdk.dart
@@ -9,10 +9,12 @@ import 'dart:typed_data';
 
 import 'src/flutter_branch_sdk_platform_interface.dart';
 import 'src/objects/app_tracking_transparency.dart';
+import 'src/objects/branch_attribution_level.dart';
 import 'src/objects/branch_universal_object.dart';
 
 export 'src/flutter_branch_sdk_platform_interface.dart';
 export 'src/objects/app_tracking_transparency.dart';
+export 'src/objects/branch_attribution_level.dart';
 export 'src/objects/branch_universal_object.dart';
 
 part 'src/flutter_branch_sdk.dart';